% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getSensorData}
\alias{getSensorData}
\title{Get sensor data from a registered senseBox of openSenseMap}
\usage{
getSensorData(senseBoxId, sensorId, startDate, endDate)
}
\arguments{
\item{senseBoxId}{a valid senseBox ID}

\item{sensorId}{a sensor ID according to the senseBox ID}

\item{startDate}{starting point for time interval as ISO8601}

\item{endDate}{end point for time interval as ISO8601}
}
\value{
Time series of a sensor as SpatialPointsDataFrame
}
\description{
Converts a time series of a senseBox sensor to a SpatialPointsDataFrame
}
\examples{
senseBox <- "5776734b230b2b1000880c0c"
sensor   <-"5776734b230b2b1000880c0f"
start    <- "2016-11-22"
end      <- "2016-11-23"
getSensorData(senseBox, sensor, start, end)

}
