% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{getSensorInfo}
\alias{getSensorInfo}
\title{Return sensor info}
\usage{
getSensorInfo(senseBoxId)
}
\arguments{
\item{senseBoxId}{a valid senseBox ID}
}
\value{
Information about all sensors as data frame
}
\description{
This returns all sensor information, in particular the sensor IDs, from a certain senseBox as data.frame
}
\examples{
getSensorInfo("56957f3ab3de1fe0052532da")

}
