% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importSenseBox}
\alias{importSenseBox}
\title{Import senseBox data from OSeM}
\usage{
importSenseBox(senseBoxId)
}
\arguments{
\item{senseBoxId}{a valid senseBox ID}
}
\value{
senseBox as spatial point in a SpatialPointsDataFrame
}
\description{
Loads a senseBox from the openSenseMap and converts it to a spatial object.
}
\examples{
importSenseBox("56957f3ab3de1fe0052532da")

}
